requirements.txt
streamlit
langchain
langchain-openai
faiss-cpu
PyPDF2
python-dotenv
